+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/Navigation.kt
@@ -98,6 +98,7 @@ internal fun ComposeActivity.Navigation() {
                     onRefresh = {
                         scope.launch {
                             (synchronizer as SdkSynchronizer).refreshAllBalances()
+                            synchronizer.refreshExchangeRateUsd()
                         }
                     }
                 )